+++ src/test/java/model/TrialCountTest.java
@@ -3,11 +3,11 @@ package model;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import common.Exception;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import response.Exception;
 
 class TrialCountTest {
   @ParameterizedTest
